"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/Module 6/m6/ADS-2Week-1/Solution.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 class PageRank {
 	Digraph dg;
 	double[] pagerank;
-	double initalpr;
+	float initalpr;
 	PageRank(Digraph digraph) {
 		this.dg = digraph;
 		pagerank = new double[dg.v()];
